// Project2/Main/src/webextension/src/interfaces/ApiResponseModel/ISendConversation.ts
export interface ISendConversation {
  fullName: string;
  alias: string;
  conversation: string;
  roleChange: string;
  cyManagerAlias: string;
  fyManagerChange: string;
  scriptLink: string;
  conversationStatus: string;
  edmValidation: string;
  fyManagerAlias: string;
  fyOrg: string;
  fyRoleSummary: string;
  fyQ1: string;
  fyQ2: string;
  fyIncentivePlan: string;
  fyCostCenter: string;
  cyOrg: string;
  cyRoleSummary: string;
  cyQ1: string;
  cyQ2: string;
  cyCareerStage: string;
  cyIncentivePlan: string;
  cyCostCenter: string;
  cyManagerFullName:string;
  hasActiveFutureManagerRequest:boolean;
  futureManagerRequestSubmittedBy:string;
  taxonomyCorrectionRequestSubmittedBy:string;
  hasActiveTaxonomyCorrectionRequest:boolean;
  fyManagerFullName:string;
  fyCareerStage:string;
  isEmployeeRecordApproved: boolean;
}